+++ apps/vmq_server/test/vmq_metrics_SUITE.erl
@@ -9,6 +9,7 @@
         ]).
 
 -export([simple_systree_test/1,
+         histogram_systree_test/1,
          simple_graphite_test/1,
          simple_prometheus_test/1,
          simple_cli_test/1,
@@ -29,18 +30,24 @@ init_per_testcase(_Case, Config) ->
     enable_on_subscribe(),
     vmq_server_cmd:set_config(allow_anonymous, true),
     vmq_server_cmd:set_config(graphite_interval, 100),
+    vmq_server_cmd:set_config(systree_interval, 100),
     vmq_server_cmd:set_config(retry_interval, 10),
+    application:set_env(vmq_server, vmq_metrics_mfa, {?MODULE, plugin_metrics, []}),
     vmq_server_cmd:listener_start(1888, []),
     vmq_metrics:reset_counters(),
     Config.
 
 end_per_testcase(_, Config) ->
+    application:unset_env(vmq_server, vmq_metrics_mfa),
+    {ok, Ret1} = vmq_server_cmd:metrics(),
+    false = lists:member({text, "counter.plugin_metrics = 123"}, Ret1),
     disable_on_subscribe(),
     vmq_test_utils:teardown(),
     Config.
 
 all() ->
     [simple_systree_test,
+     histogram_systree_test,
      simple_graphite_test,
      simple_prometheus_test,
      simple_cli_test,
@@ -53,17 +60,33 @@ simple_systree_test(_) ->
     ok = packet:expect_packet(Socket, "publish", Publish),
     gen_tcp:close(Socket).
 
+histogram_systree_test(_) ->
+    histogram_systree_test("/count", 10),
+    histogram_systree_test("/sum", 100),
+    histogram_systree_test("/bucket/10", 4),
+    histogram_systree_test("/bucket/100", 6),
+    histogram_systree_test("/bucket/1000", 8),
+    histogram_systree_test("/bucket/inf", 10).
+
 simple_graphite_test(_) ->
     vmq_server_cmd:set_config(graphite_enabled, true),
     vmq_server_cmd:set_config(graphite_host, "localhost"),
     vmq_server_cmd:set_config(graphite_interval, 1000),
     vmq_server_cmd:set_config(graphite_include_labels, false),
+    % ensure we have a subscription that we can count later on
     SubSocket = sample_subscribe(),
     {ok, LSocket} = gen_tcp:listen(2003, [binary, {packet, raw},
                                           {active, false},
                                           {reuseaddr, true}]),
     {ok, GraphiteSocket1} = gen_tcp:accept(LSocket), %% vmq_graphite connects
-    Want = [<<"mqtt.subscribe.received 1">>],
+    Want = [<<"mqtt.subscribe.received 1">>
+            ,<<"plugin.histogram.count 10">>
+            ,<<"plugin.histogram.sum 100">>
+            ,<<"plugin.histogram.bucket.10 4">>
+            ,<<"plugin.histogram.bucket.100 6">>
+            ,<<"plugin.histogram.bucket.1000 8">>
+            ,<<"plugin.histogram.bucket.inf 10">>
+           ],
     true = recv_data(GraphiteSocket1, Want),
     gen_tcp:close(GraphiteSocket1),
 
@@ -96,12 +119,12 @@ recv_data(Socket, Want0) ->
                       re:split(Data0, "\n")),
             Want1 =
             lists:foldl(
-              fun(Line, WantAcc) ->
-                      case lists:member(Line, WantAcc) of
-                          true -> WantAcc -- [Line];
+              fun(WItem, WantAcc) ->
+                      case lists:member(WItem, Data1) of
+                          true -> WantAcc -- [WItem];
                           false -> WantAcc
                       end
-              end, Want0, Data1),
+              end, Want0, Want0),
             case Want1 of
                 [] -> true;
                 _ ->
@@ -120,12 +143,29 @@ simple_prometheus_test(_) ->
     {ok, {_Status, _Headers, Body}} = httpc:request("http://localhost:8888/metrics"),
     Lines = re:split(Body, "\n"),
     Node = atom_to_list(node()),
-    Line = list_to_binary(
-             "mqtt_subscribe_received{node=\"" ++ Node ++ "\",mqtt_version=\"4\"} 1"),
-    true = lists:foldl(fun(L, _) when L == Line ->
-                               true;
-                          (_, Acc) -> Acc
-                       end, false, Lines),
+    true = lists:member(
+             list_to_binary(
+               "mqtt_subscribe_received{node=\"" ++ Node ++ "\",mqtt_version=\"4\"} 1"), Lines),
+
+    true = lists:member(
+             list_to_binary(
+               "plugin_histogram_count{node=\"" ++ Node ++ "\"} 10"), Lines),
+    true = lists:member(
+             list_to_binary(
+               "plugin_histogram_sum{node=\"" ++ Node ++ "\"} 100"), Lines),
+    true = lists:member(
+             list_to_binary(
+               "plugin_histogram_bucket{node=\"" ++ Node ++ "\",le=\"10\"} 4"), Lines),
+    true = lists:member(
+             list_to_binary(
+               "plugin_histogram_bucket{node=\"" ++ Node ++ "\",le=\"100\"} 6"), Lines),
+    true = lists:member(
+             list_to_binary(
+               "plugin_histogram_bucket{node=\"" ++ Node ++ "\",le=\"1000\"} 8"), Lines),
+    true = lists:member(
+             list_to_binary(
+               "plugin_histogram_bucket{node=\"" ++ Node ++ "\",le=\"+Inf\"} 10"), Lines),
+
     gen_tcp:close(SubSocket).
 
 simple_cli_test(_) ->
@@ -138,21 +178,20 @@ simple_cli_test(_) ->
     gen_tcp:close(SubSocket).
 
 pluggable_metrics_test(_) ->
-    application:set_env(vmq_server, vmq_metrics_mfa, {?MODULE, plugin_metrics, []}),
     {ok, Ret0} = vmq_server_cmd:metrics(),
-    true =
-    lists:foldl(fun({text, "counter.plugin_metrics = 123"}, _) -> true;
-                   (_, Acc) -> Acc
-                end, false, Ret0),
-    application:unset_env(vmq_server, vmq_metrics_mfa),
-    {ok, Ret1} = vmq_server_cmd:metrics(),
-    true =
-    lists:foldl(fun({text, "counter.plugin_metrics = 123"}, _) -> false;
-                   (_, Acc) -> Acc
-                end, true, Ret1).
+
+    true = lists:member({text, "counter.plugin_metrics = 123"}, Ret0),
+    true = lists:member({text, "histogram.plugin_histogram_count = 10"}, Ret0),
+    true = lists:member({text, "histogram.plugin_histogram_sum = 100"}, Ret0),
+    true = lists:member({text, "histogram.plugin_histogram_bucket_10 = 4"}, Ret0),
+    true = lists:member({text, "histogram.plugin_histogram_bucket_100 = 6"}, Ret0),
+    true = lists:member({text, "histogram.plugin_histogram_bucket_1000 = 8"}, Ret0),
+    true = lists:member({text, "histogram.plugin_histogram_bucket_infinity = 10"}, Ret0).
 
 plugin_metrics() ->
-    [{counter, [], plugin_metrics, plugin_metrics, <<"Simple Plugin Metric">>, 123}].
+    [{counter, [], plugin_metrics, plugin_metrics, <<"Simple Plugin Metric">>, 123},
+     {histogram, [], plugin_histogram, plugin_histogram, <<"Simple Plugin Histogram">>,
+      {10, 100, #{10 => 4,  100 => 6, 1000 => 8, infinity => 10}}}].
 
 enable_on_subscribe() ->
     vmq_plugin_mgr:enable_module_plugin(
@@ -173,3 +212,19 @@ sample_subscribe() ->
     ok = gen_tcp:send(SubSocket, Subscribe),
     ok = packet:expect_packet(SubSocket, "suback", Suback),
     SubSocket.
+
+histogram_systree_test(Suffix, Val) ->
+    %% let the metrics system do some increments
+    Connect = packet:gen_connect("hist-test", [{keepalive,60}]),
+    Connack = packet:gen_connack(0),
+    Topic =  "$SYS/+/plugin/histogram" ++ Suffix,
+    Subscribe = packet:gen_subscribe(53, Topic, 0),
+    Suback = packet:gen_suback(53, 0),
+    {ok, SubSocket} = packet:do_client_connect(Connect, Connack, []),
+    ok = gen_tcp:send(SubSocket, Subscribe),
+    ok = packet:expect_packet(SubSocket, "suback", Suback),
+    PublishTopic = "$SYS/" ++ atom_to_list(node()) ++ "/plugin/histogram" ++ Suffix,
+    Publish = packet:gen_publish(PublishTopic, 0, integer_to_binary(Val), []),
+    ok = packet:expect_packet(SubSocket, "publish", Publish),
+    gen_tcp:send(SubSocket, packet:gen_disconnect()),
+    gen_tcp:close(SubSocket).