+++ changelog.md
@@ -45,6 +45,7 @@
 - Optimization: optimize away read operation during client registration by
   reusing already available data.
 - Enable plugins to expose metrics via the available metric providers.
+- Add histogram metric type.
 
 ## VerneMQ 1.6.0